Abstracted out the ArgumentReachesValue class for use in the LengthAnnotator later.
